Bugfixes for python 3 compatibility and reproducibility.
Thoses changes ensure the compatibility with python 2 and python 3.
They also assure near reproducibility between run in python 2 and python 3.
This imply that some change in order are to expect, specially in commentary but they should be marginal. This is linked to the unordered dict in python3 (in this version, nearly all dictionnaries has been replaced by OrderedDict from the collections package).